current client (nonfunctional)
