+++ public/app.js
@@ -3,8 +3,9 @@ import { global } from "./lib/global.js";
 import { config } from "./lib/config.js";
 import { Canvas } from "./lib/canvas.js";
 import { color } from "./lib/color.js";
+import { gameDraw } from "./lib/gameDraw.js";
 import * as socketStuff from "./lib/socketInit.js";
-(async function (util, global, config, Canvas, color, socketStuff) {
+(async function (util, global, config, Canvas, color, gameDraw, socketStuff) {
 
 let { socketInit, gui, leaderboard, minimap, moveCompensation, lag, getNow } = socketStuff;
 // fetch("changelog.md", { cache: "no-cache" })
@@ -80,458 +81,7 @@ let animations = window.animations = {
     minimap: new Animation(-1, 1, 0.025),
     leaderboard: new Animation(-1, 1, 0.025)
 };
-// Color functions
-/** https://gist.github.com/jedfoster/7939513 **/
-function decimal2hex(d) {
-    return d.toString(16);
-} // convert a decimal value to hex
-function hex2decimal(h) {
-    return parseInt(h, 16);
-} // convert a hex value to decimal
-let mixColors = (color_2, color_1, weight = 0.5) => {
-    if (weight === 1) return color_1;
-    if (weight === 0) return color_2;
-    var col = "#";
-    for (var i = 1; i <= 6; i += 2) {
-        // loop through each of the 3 hex pairs—red, green, and blue, skip the '#'
-        var v1 = hex2decimal(color_1.substr(i, 2)), // extract the current pairs
-            v2 = hex2decimal(color_2.substr(i, 2)),
-            // combine the current pairs from each source color, according to the specified weight
-            val = decimal2hex(Math.floor(v2 + (v1 - v2) * weight));
-        while (val.length < 2) {
-            val = "0" + val;
-        } // prepend a '0' if val results in a single digit
-        col += val; // concatenate val to our new color string
-    }
-    return col; // PROFIT!
-};
-function hslToRgb(h, s, l) {
-    let r, g, b;
 
-    if (s === 0) {
-        r = g = b = l; // achromatic
-    } else {
-        const q = l < 0.5 ? l * (1 + s) : l + s - l * s;
-        const p = 2 * l - q;
-        r = hueToRgb(p, q, h + 1/3);
-        g = hueToRgb(p, q, h);
-        b = hueToRgb(p, q, h - 1/3);
-    }
-    return '#' +
-        Math.round(r * 255).toString(16).padStart(2, '0') +
-        Math.round(g * 255).toString(16).padStart(2, '0') +
-        Math.round(b * 255).toString(16).padStart(2, '0');
-}
-function rgbToHsl(rgb) {
-    let r, g, b, h, s, l;
-
-    r = parseInt(rgb.substring(1, 3), 16) / 255;
-    g = parseInt(rgb.substring(3, 5), 16) / 255;
-    b = parseInt(rgb.substring(5, 7), 16) / 255;
-
-    let cmax = Math.max(r, g, b);
-    let cmin = Math.min(r, g, b);
-    let deltaC = cmax - cmin;
-
-    // Hue
-    switch (true){
-        case deltaC == 0:
-            h = 0;
-            break;
-        case cmax == r:
-            h = 1/6 * (((g - b) / deltaC) % 6);
-            break;
-        case cmax == g:
-            h = 1/6 * ((b - r) / deltaC + 2);
-            break;
-        case cmax == b:
-            h = 1/6 * ((r - g) / deltaC + 4);
-            break;
-    }
-    // Brightness
-    l = (cmax + cmin) / 2
-    // Saturation
-    if (deltaC == 0)
-        s = 0;
-    else 
-        s = deltaC / (1 - Math.abs(2 * l - 1));
-
-    return [h, s, l];
-}
-function hueToRgb(p, q, t) {
-    if (t < 0) t += 1;
-    if (t > 1) t -= 1;
-    if (t < 0.166) return p + (q - p) * 6 * t;
-    if (t < 0.5  ) return q;
-    if (t < 0.666) return p + (q - p) * (2/3 - t) * 6;
-    return p;
-}
-function clamp(n, lower, upper) {
-    return Math.min(upper, Math.max(lower, n));
-}
-//TODO: somehow move the calculation to these in reanimateColors to improve performance
-var colorCache = {};
-function modifyColor(color, base = "16 0 1 0 false") {
-    // Edge cases because spaghetti
-    if (typeof color == 'number') {
-        color = color + " 0 1 0 false";
-    }
-    if (typeof base == 'number') {
-        base = base + " 0 1 0 false";
-    }
-    // Split into array
-    let colorDetails = color.split(" "),
-        baseDetails = base.split(" ");
-
-    // Color mirroring
-    if (colorDetails[0] == "-1") {
-        colorDetails[0] = baseDetails[0];
-    }
-    if (colorDetails[0] == "-1") {
-        colorDetails[0] = gui.color;
-    }
-
-    let colorId = "";
-    for (let i in colorDetails) {
-        colorId += colorDetails[i] + " ";
-    }
-
-    // Exit if calculated already
-    let cachedColor = colorCache[colorId];
-    if (cachedColor != undefined) return cachedColor;
-
-    // Get HSL values
-    let baseColor = colorDetails[0];
-    // check if color.base is not a word.
-    if (!isNaN(baseColor)) {
-        baseColor = parseInt(baseColor);
-    }
-    baseColor = rgbToHsl(getColor(baseColor) ?? baseColor);
-    
-    // Get color config
-    let hueShift = parseFloat(colorDetails[1]) / 360,
-        saturationShift = parseFloat(colorDetails[2]),
-        brightnessShift = parseFloat(colorDetails[3]) / 100,
-        allowBrightnessInvert = colorDetails[4] == 'true';
-
-    // Apply config
-    let finalHue = (baseColor[0] + hueShift) % 1,
-        finalSaturation = clamp(baseColor[1] * saturationShift, 0, 1),
-        finalBrightness = baseColor[2] + brightnessShift;
-
-    if (allowBrightnessInvert && (finalBrightness > 1 || finalBrightness < 0)) {
-        finalBrightness -= brightnessShift * 2;
-    }
-    finalBrightness = clamp(finalBrightness, 0, 1);
-
-    // Gaming.
-    let finalColor = hslToRgb(finalHue, finalSaturation, finalBrightness);
-    if (!animatedColors[colorDetails[0]]) colorCache[colorId] = finalColor
-    return finalColor;
-}
-function getRainbow(a, b, c = 0.5) {
-    if (0 >= c) return a;
-    if (1 <= c) return b;
-    let f = 1 - c;
-    a = parseInt(a.slice(1, 7), 16);
-    b = parseInt(b.slice(1, 7), 16);
-    return (
-        "#" +
-        (
-            (((a & 0xff0000) * f + (b & 0xff0000) * c) & 0xff0000) |
-            (((a & 0x00ff00) * f + (b & 0x00ff00) * c) & 0x00ff00) |
-            (((a & 0x0000ff) * f + (b & 0x0000ff) * c) & 0x0000ff)
-        )
-            .toString(16)
-            .padStart(6, "0")
-    );
-}
-let animatedColor = {
-    lesbian: "",
-    gay: "",
-    bi: "",
-    trans: "",
-    blue_red: "",
-    blue_grey: "",
-    grey_blue: "",
-    red_grey: "",
-    grey_red: ""
-};
-function reanimateColors() {
-    let now = Date.now(),
-
-        //six_gradient = Math.floor((now / 200) % 6),
-        five_bars = Math.floor((now % 2000) / 400),
-        three_bars = Math.floor((now % 2000) * 3 / 2000),
-        blinker = 150 > now % 300,
-
-        lesbian_magenta  = "#a50062",
-        lesbian_oredange = "#d62900",
-        lesbian_white    = "#ffffff",
-        lesbian_useSecondSet = five_bars < 2,
-
-        gay_transition = (now / 2000) % 1,
-
-        bi_pink   = "#D70071",
-        bi_purple = "#9C4E97",
-        bi_blue   = "#0035AA",
-
-        trans_pink  = "#f7a8b8",
-        trans_blue  = "#55cdfc",
-        trans_white = "#ffffff";
-
-    animatedColor.lesbian = getRainbow(lesbian_useSecondSet ? lesbian_oredange : lesbian_white, lesbian_useSecondSet ? lesbian_white : lesbian_magenta, (lesbian_useSecondSet ? five_bars : five_bars - 3) / 2);
-    animatedColor.gay = hslToRgb(gay_transition, 0.75, 0.5);
-    animatedColor.bi = [bi_pink, bi_purple, bi_blue][three_bars];
-    animatedColor.trans = [trans_blue, trans_pink, trans_white, trans_pink, trans_blue][five_bars];
-
-    animatedColor.blue_red = blinker ? color.blue : color.red;
-    animatedColor.blue_grey = blinker ? color.blue : color.grey;
-    animatedColor.grey_blue = blinker ? color.grey : color.blue;
-    animatedColor.red_grey = blinker ? color.red : color.grey;
-    animatedColor.grey_red = blinker ? color.grey : color.red;
-}
-const animatedColors = {
-    // police
-    20: true,
-    animatedBlueRed: true,
-
-    21: true,
-    animatedBlueGrey: true,
-    animatedBlueGray: true,
-
-    22: true,
-    animatedGreyBlue: true,
-    animatedGrayBlue: true,
-
-    23: true,
-    animatedRedGrey: true,
-    animatedRedGray: true,
-
-    24: true,
-    animatedGreyRed: true,
-    animatedGrayRed: true,
-
-    // lesbian
-    29: true,
-    animatedLesbian: true,
-
-    // rainbow
-    rainbow: true,
-
-    // trans
-    37: true,
-    animatedTrans: true,
-
-    38: true,
-    animatedBi: true,
-}
-function getColor(colorNumber) {
-    switch (colorNumber) {
-        case 0:
-        case "teal":
-        case "aqua":
-            return color.teal;
-        case 1:
-        case "lightGreen":
-            return color.lgreen;
-        case 2:
-        case "orange":
-            return color.orange;
-        case 3:
-        case "yellow":
-            return color.yellow;
-        case 4:
-        case "lavender":
-            return color.lavender;
-        case 5:
-        case "pink":
-            return color.pink;
-        case 6:
-        case "veryLightGrey":
-        case "veryLightGray":
-            return color.vlgrey;
-        case 7:
-        case "lightGrey":
-        case "lightGray":
-            return color.lgrey;
-        case 8:
-        case "pureWhite":
-            return color.guiwhite;
-        case 9:
-        case "black":
-            return color.black;
-        case 10:
-        case "blue":
-            return color.blue;
-        case 11:
-        case "green":
-            return color.green;
-        case 12:
-        case "red":
-            return color.red;
-        case 13:
-        case "gold":
-            return color.gold;
-        case 14:
-        case "purple":
-            return color.purple;
-        case 15:
-        case "magenta":
-            return color.magenta;
-        case 16:
-        case "grey":
-        case "gray":
-            return color.grey;
-        case 17:
-        case "darkGrey":
-        case "darkGray":
-            return color.dgrey;
-        case 18:
-        case "white":
-            return color.white;
-        case 19:
-        case "pureBlack":
-            return color.guiblack;
-        case 20:
-        case "animatedBlueRed":
-            return animatedColor.blue_red;
-        case 21:
-        case "animatedBlueGrey":
-        case "animatedBlueGray":
-            return animatedColor.blue_grey;
-        case 22:
-        case "animatedGreyBlue":
-        case "animatedGrayBlue":
-            return animatedColor.grey_blue;
-        case 23:
-        case "animatedRedGrey":
-        case "animatedRedGray":
-            return animatedColor.red_grey;
-        case 24:
-        case "animatedGreyRed":
-        case "animatedGrayRed":
-            return animatedColor.grey_red;
-        case 25:
-        case "mustard":
-            return "#C49608";
-        case 26:
-        case "darkOrange":
-            return "#EC7B0F";
-        case 27:
-        case "brown":
-            return "#895918";
-        case 28:
-        case "cyan":
-        case "turquoise":
-            return "#13808E";
-        case 29:
-        case "animatedLesbian":
-            return animatedColor.lesbian;
-        case 30:
-        case "powerGem":
-        case "powerStone":
-            return "#a913cf";
-        case 31:
-        case "spaceGem":
-        case "spaceStone":
-            return "#226ef6";
-        case 32:
-        case "realityGem":
-        case "realityStone":
-            return "#ff1000";
-        case 33:
-        case "soulGem":
-        case "soulStone":
-            return "#ff9000";
-        case 34:
-        case "timeGem":
-        case "timeStone":
-            return "#00e00b";
-        case 35:
-        case "mindGem":
-        case "mindStone":
-            return "#ffd300";
-        case 36:
-        case "rainbow":
-            return animatedColor.gay;
-        case 37:
-        case "animatedTrans":
-            return animatedColor.trans;
-        case 38:
-        case "animatedBi":
-            return animatedColor.bi;
-        case 39:
-        case "pumpkinStem":
-            return "#654321";
-        case 40:
-        case "pumpkinBody":
-            return "#e58100";
-        case 41:
-        case "tree":
-            return "#267524";
-    }
-}
-function getColorDark(givenColor) {
-    let dark = config.graphical.neon ? color.white : color.black;
-    if (config.graphical.darkBorders) return dark;
-    return mixColors(givenColor, dark, color.border);
-}
-function getZoneColor(cell, real) {
-    switch (cell) {
-        case "bas1":
-        case "bap1":
-        case "dom1":
-            return color.blue;
-        case "bas2":
-        case "bap2":
-        case "dom2":
-            return color.green;
-        case "bas3":
-        case "bap3":
-        case "dom3":
-        case "boss":
-            return color.red;
-        case "bas4":
-        case "bap4":
-        case "dom4":
-            return color.magenta;
-        case "bas5":
-        case "bap5":
-        case "dom5":
-            return "#C49608";
-        case "bas6":
-        case "bap6":
-        case "dom6":
-            return "#EC7B0F";
-        case "bas7":
-        case "bap7":
-        case "dom7":
-            return "#895918";
-        case "bas8":
-        case "bap8":
-        case "dom8":
-            return "#13808E";
-        case "port":
-            return color.guiblack;
-        case "nest":
-            return real ? color.purple : color.lavender;
-        case "dom0":
-            return color.gold;
-        default:
-            return real ? color.white : color.lgrey;
-    }
-}
-
-function setColor(context, givenColor) {
-    if (config.graphical.neon) {
-        context.fillStyle = getColorDark(givenColor);
-        context.strokeStyle = givenColor;
-    } else {
-        context.fillStyle = givenColor;
-        context.strokeStyle = getColorDark(givenColor);
-    }
-}
 // Mockup functions
 // Prepare stuff
 global.player = {
@@ -872,7 +422,7 @@ function drawText(rawText, x, y, size, defaultFillStyle, align = "left", center 
                 if (!isNaN(str)) {
                     str = parseInt(str);
                 }
-                str = getColor(str) ?? str;
+                str = gameDraw.getColor(str) ?? str;
             }
             context.fillStyle = str;
 
@@ -1076,16 +626,16 @@ const drawEntity = (baseColor, x, y, instance, ratio, alpha = 1, scale = 1, rot 
             let position = positions[i] / (g.aspect === 1 ? 2 : 1),
                 gx = g.offset * Math.cos(g.direction + g.angle + rot) + (g.length / 2 - position) * Math.cos(g.angle + rot),
                 gy = g.offset * Math.sin(g.direction + g.angle + rot) + (g.length / 2 - position) * Math.sin(g.angle + rot),
-                gunColor = g.color == null ? color.grey : modifyColor(g.color, baseColor),
+                gunColor = g.color == null ? color.grey : gameDraw.modifyColor(g.color, baseColor),
                 borderless = g.borderless,
                 fill = g.drawFill;
-            setColor(context, mixColors(gunColor, render.status.getColor(), render.status.getBlend()));
+            gameDraw.setColor(context, gameDraw.mixColors(gunColor, render.status.getColor(), render.status.getBlend()));
             drawTrapezoid(context, xx + drawSize * gx, yy + drawSize * gy, drawSize * (g.length / 2 - (g.aspect === 1 ? position * 2 : 0)), (drawSize * g.width) / 2, g.aspect, g.angle + rot, borderless, fill);
         }
     }
     // Draw body
     context.globalAlpha = 1;
-    setColor(context, mixColors(modifyColor(instance.color, baseColor), render.status.getColor(), render.status.getBlend()));
+    gameDraw.setColor(context, gameDraw.mixColors(gameDraw.modifyColor(instance.color, baseColor), render.status.getColor(), render.status.getBlend()));
     drawPoly(context, xx, yy, (drawSize / m.size) * m.realSize, m.shape, rot, m.borderless, m.drawFill);
     // Draw guns above us
     context.lineWidth = Math.max(config.graphical.mininumBorderChunk, ratio * config.graphical.borderChunk);
@@ -1095,10 +645,10 @@ const drawEntity = (baseColor, x, y, instance, ratio, alpha = 1, scale = 1, rot 
             let position = positions[i] / (g.aspect === 1 ? 2 : 1),
                 gx = g.offset * Math.cos(g.direction + g.angle + rot) + (g.length / 2 - position) * Math.cos(g.angle + rot),
                 gy = g.offset * Math.sin(g.direction + g.angle + rot) + (g.length / 2 - position) * Math.sin(g.angle + rot),
-                gunColor = g.color == null ? color.grey : modifyColor(g.color, baseColor),
+                gunColor = g.color == null ? color.grey : gameDraw.modifyColor(g.color, baseColor),
                 borderless = g.borderless,
                 fill = g.drawFill;
-            setColor(context, mixColors(gunColor, render.status.getColor(), render.status.getBlend()));
+            gameDraw.setColor(context, gameDraw.mixColors(gunColor, render.status.getColor(), render.status.getBlend()));
             drawTrapezoid(context, xx + drawSize * gx, yy + drawSize * gy, drawSize * (g.length / 2 - (g.aspect === 1 ? position * 2 : 0)), (drawSize * g.width) / 2, g.aspect, g.angle + rot, borderless, fill);
         }
     }
@@ -1138,7 +688,7 @@ function drawHealth(x, y, instance, ratio, alpha) {
         let health = instance.render.health.get(),
             shield = instance.render.shield.get();
         if (health < 1 - 1e-4 || shield < 1 - 1e-4) {
-            const col = config.graphical.coloredHealthbars ? mixColors(getColor(instance.color), color.guiwhite, 0.5) : color.lgreen;
+            const col = config.graphical.coloredHealthbars ? gameDraw.mixColors(gameDraw.getColor(instance.color), color.guiwhite, 0.5) : color.lgreen;
             let yy = y + 1.1 * realSize + 15;
             let barWidth = 5;
             ctx.globalAlpha = alpha * alpha * fade;
@@ -1148,7 +698,7 @@ function drawHealth(x, y, instance, ratio, alpha) {
             drawBar(x - size, x - size + 2 * size * health, yy + barWidth * config.graphical.seperatedHealthbars, barWidth, col);
             if (shield || config.graphical.seperatedHealthbars) {
                 if (!config.graphical.seperatedHealthbars) ctx.globalAlpha = (0.3 + shield * 0.3) * alpha * alpha * fade;
-                drawBar(x - size, x - size + 2 * size * shield, yy, barWidth, config.graphical.coloredHealthbars ? mixColors(col, color.guiblack, 0.25) : color.teal);
+                drawBar(x - size, x - size + 2 * size * shield, yy, barWidth, config.graphical.coloredHealthbars ? gameDraw.mixColors(col, color.guiblack, 0.25) : color.teal);
                 ctx.globalAlpha = 1;
             }
         }
@@ -1360,7 +910,7 @@ function drawFloor(px, py, ratio) {
             ctx.fillStyle = config.graphical.screenshotMode ? color.guiwhite : color.white;
             ctx.fillRect(left, top, right - left, bottom - top);
             ctx.globalAlpha = 0.3;
-            ctx.fillStyle = config.graphical.screenshotMode ? color.guiwhite : getZoneColor(tile, true);
+            ctx.fillStyle = config.graphical.screenshotMode ? color.guiwhite : gameDraw.getZoneColor(tile, true);
             ctx.fillRect(left, top, right - left, bottom - top);
         }
     }
@@ -1436,7 +986,7 @@ function drawEntities(px, py, ratio) {
                 alpha = Math.max(0, Math.min(1000, chat.expires - now) / 1000);
 
             ctx.globalAlpha = 0.5 * alpha;
-            drawBar(x - msgLength / 2, x + msgLength / 2, y, 30, modifyColor(instance.color));
+            drawBar(x - msgLength / 2, x + msgLength / 2, y, 30, gameDraw.modifyColor(instance.color));
             ctx.globalAlpha = alpha;
             config.graphical.fontStrokeRatio *= 1.2;
             drawText(text, x, y + 7, 15, color.guiwhite, "center");
@@ -1497,10 +1047,10 @@ function drawUpgradeTree() {
             size = tileSize;
         if (ax < -50 || ax + size - 50 > global.screenWidth) continue;
         ctx.globalAlpha = 0.75;
-        ctx.fillStyle = getColor(10);
+        ctx.fillStyle = gameDraw.getColor(10);
         drawGuiRect(ax, ay, size, size);
         ctx.globalAlpha = 0.15;
-        ctx.fillStyle = getColor(0);
+        ctx.fillStyle = gameDraw.getColor(0);
         drawGuiRect(ax, ay, size, size * 0.6);
         ctx.fillStyle = color.black;
         drawGuiRect(ax, ay + size * 0.6, size, size * 0.4);
@@ -1707,8 +1257,8 @@ function drawMinimapAndDebug(spacing, alcoveSize) {
         let j = 0;
         for (let xcell = 0; xcell < W; xcell++) {
             let cell = global.roomSetup[ycell][xcell];
-            ctx.fillStyle = getZoneColor(cell);
-            if (getZoneColor(cell) !== color.white) {
+            ctx.fillStyle = gameDraw.getZoneColor(cell);
+            if (gameDraw.getZoneColor(cell) !== color.white) {
                 drawGuiRect(x + (j * len) / W, y + (i * height) / H, len / W, height / H);
             }
             j++;
@@ -1722,7 +1272,7 @@ function drawMinimapAndDebug(spacing, alcoveSize) {
     ctx.fillStyle = color.black;
     drawGuiRect(x, y, len, height, true); // Border
     for (let entity of minimap.get()) {
-        ctx.fillStyle = mixColors(modifyColor(entity.color), color.black, 0.3);
+        ctx.fillStyle = gameDraw.mixColors(gameDraw.modifyColor(entity.color), color.black, 0.3);
         ctx.globalAlpha = entity.alpha;
         switch (entity.type) {
             case 2:
@@ -1780,7 +1330,7 @@ function drawLeaderboard(spacing, alcoveSize, max) {
         drawBar(x, x + len, y + height / 2, height - 3 + config.graphical.barChunk, color.black);
         drawBar(x, x + len, y + height / 2, height - 3, color.grey);
         let shift = Math.min(1, entry.score / max);
-        drawBar(x, x + len * shift, y + height / 2, height - 3.5, modifyColor(entry.barColor));
+        drawBar(x, x + len * shift, y + height / 2, height - 3.5, gameDraw.modifyColor(entry.barColor));
         // Leadboard name + score
         let nameColor = entry.nameColor || "#FFFFFF";
         drawText(entry.label + (": " + util.handleLargeNumber(Math.round(entry.score))), x + len / 2, y + height / 2, height - 5, nameColor, "center", true);
@@ -1824,10 +1374,10 @@ function drawAvailableUpgrades(spacing, alcoveSize) {
 
             // Draw box
             ctx.globalAlpha = 0.5;
-            ctx.fillStyle = getColor(colorIndex > 18 ? colorIndex - 19 : colorIndex);
+            ctx.fillStyle = gameDraw.getColor(colorIndex > 18 ? colorIndex - 19 : colorIndex);
             drawGuiRect(x, y, len, height);
             ctx.globalAlpha = 0.1;
-            ctx.fillStyle = getColor(-10 + colorIndex++);
+            ctx.fillStyle = gameDraw.getColor(-10 + colorIndex++);
             drawGuiRect(x, y, len, height * 0.6);
             ctx.fillStyle = color.black;
             drawGuiRect(x, y + height * 0.6, len, height * 0.4);
@@ -1881,7 +1431,7 @@ function drawAvailableUpgrades(spacing, alcoveSize) {
     }
 }
 
-const gameDraw = (ratio, drawRatio) => {
+const gameDrawAlive = (ratio, drawRatio) => {
     let GRAPHDATA = 0;
     // Prep stuff
     renderTimes++;
@@ -2017,7 +1567,7 @@ const gameDrawDisconnected = () => {
         if (!unset) ratio *= by;
     };
     scaleScreenRatio(ratio, true);
-    clearScreen(mixColors(color.red, color.guiblack, 0.3), 0.25);
+    clearScreen(gameDraw.mixColors(color.red, color.guiblack, 0.3), 0.25);
     let shift = animations.disconnected.get();
     ctx.translate(0, -shift * global.screenHeight);
     drawText("Disconnected", global.screenWidth / 2, global.screenHeight / 2, 30, color.guiwhite, "center");
@@ -2027,7 +1577,7 @@ const gameDrawDisconnected = () => {
 // The main function
 function animloop() {
     global.animLoopHandle = window.requestAnimFrame(animloop);
-    reanimateColors();
+    gameDraw.reanimateColors();
     global.player.renderv += (global.player.view - global.player.renderv) / 30;
     var ratio = config.graphical.screenshotMode ? 2 : util.getRatio();
     // Set the drawing style
@@ -2052,7 +1602,7 @@ function animloop() {
     }
     ctx.translate(0.5, 0.5);
     if (global.gameStart) {
-        gameDraw(ratio, util.getScreenRatio());
+        gameDrawAlive(ratio, util.getScreenRatio());
     } else if (!global.disconnected) {
         gameDrawBeforeStart();
     }
@@ -2065,4 +1615,4 @@ function animloop() {
     ctx.translate(-0.5, -0.5);
 }
 
-})(util, global, config, Canvas, color, socketStuff);
+})(util, global, config, Canvas, color, gameDraw, socketStuff);
