+++ public/index.html
@@ -122,6 +122,7 @@
     <script type="module" src="./lib/global.js"></script>
     <script type="module" src="./lib/util.js"></script>
     <script type="module" src="./lib/canvas.js"></script>
+    <script type="module" src="./lib/gameDraw.js"></script>
     <script type="module" src="./lib/protocol.js"></script>
     <script type="module" src="./lib/socketInit.js"></script>
     <script type="module" src="./app.js"></script>
